rename `object-tree.ts` to `types.object-tree.d.ts`
