+++ packages/plugins/io-center/src/ui/components/object-tree/types.object-tree.d.ts
@@ -1,4 +1,3 @@
-
 export type ObjectTree = TreeNode[]
 
 export type TreeNode = {
